yeahbaby

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/model/Accounts.java
@@ -1,13 +1,15 @@
 package com.techelevator.tenmo.model;
 
+import java.math.BigDecimal;
+
 public class Accounts {
 	private int account_id;
 	private int user_id;
-	private double balance;
+	private BigDecimal balance;
 	
 	public Accounts() {}
 	
-	public Accounts(int accountId, int userId, double balance) {
+	public Accounts(int accountId, int userId, BigDecimal balance) {
 		this.account_id = accountId;
 		this.user_id = userId;
 		this.balance = balance;
@@ -29,11 +31,11 @@ public class Accounts {
 		this.user_id = user_id;
 	}
 
-	public double getBalance() {
+	public BigDecimal getBalance() {
 		return balance;
 	}
 
-	public void setBalance(double balance) {
+	public void setBalance(BigDecimal balance) {
 		this.balance = balance;
 	}
